Address PR reviews and refactor
