+++ langchain4j-bom/pom.xml
@@ -598,7 +598,7 @@
             <dependency>
                 <groupId>dev.langchain4j</groupId>
                 <artifactId>langchain4j-observation</artifactId>
-                <version>${project.version}</version>
+                <version>${langchain4j.beta.version}</version>
             </dependency>
 
             <!-- spring boot -->
