+++ langchain4j-observation/pom.xml
@@ -33,7 +33,6 @@
             <groupId>dev.langchain4j</groupId>
             <artifactId>langchain4j-core</artifactId>
             <version>1.12.0-SNAPSHOT</version>
-            <scope>compile</scope>
         </dependency>
         <dependency>
             <groupId>io.micrometer</groupId>
@@ -56,6 +55,7 @@
             <groupId>org.awaitility</groupId>
             <artifactId>awaitility</artifactId>
             <version>4.3.0</version>
+            <scope>test</scope>
         </dependency>
     </dependencies>
 </project>
