+++ langchain4j-observation/src/main/java/context/ChatModelObservationContext.java
@@ -1,4 +1,4 @@
-package dev.langchain4j.observation.listeners;
+package context;
 
 import dev.langchain4j.model.chat.listener.ChatModelErrorContext;
 import dev.langchain4j.model.chat.listener.ChatModelRequestContext;
